Add a script for supplementing historic benchmark results manually
If you have started to collect benchmark results at a certain stage of your project but now also would like to compare these results to older historic values, the `run_local` file can be run to check in these historic data.

This patch also adds a section to the docs that explains the new script.