Enable user to specify the ref being tested
The ref will be used preferentially (if specified) when looking up the details
of the commit to record the results against.

Why:
When used to record results from a performance test run triggered by a
comment on a PR, the results were recorded against the head commit of main
on our branch. Not the tested commit. This change allows users to specify
the tested ref.

Example usage:
```
with:
  ref: refs/pull/${{ github.event.issue.number }}/head
```