Add support for Nyrkiö change detection
Adds optional support for using nyrkio.com change
detection service. Consequently gh-pages-branch and
benchmark-data-dir-path are made optional parameters
and if not given, the original method of storing results
in github and calculating a threshold based alert is not
executed.

On the other hand if nyrkio-enabled is false (the default),
Nyrkiö is not used.

(Note: it's possible but probably not useful to enable both as
this will just lead to more false positives, if anything.)

In addition, this commit adds a new tool "criterion". There already
was some kind of criterion support under the "cargo" tool, but it
didn't work for the output.txt file I got from turso/limbo project.
I don't have a lot of insight into what is the typical way of running
criterion, so understand if this particular case is not representative.

This implementation will work for post-commit runs. The Nyrkiö APIs
for pull requests aren't used in this commit and you will probably
create a mess if you enable this for pull requests.

As usual, didn't add any testing but added my test file(s) under
test/data.

To be able to develop and test locally I added support for local
git repo (which must be CWD). In fact, the branch name is often
picked up from the local filesystem git repo. It is required in
the Nyrkiö API but often absent in GitHub API objects.