Challenge Publish Handshake
The repo secrets aren't available to PR authors that aren't part
of the owner org or otherwise have write access to the repo.

To enable such 3rd party contributors to still benefit from Nyrkiö
running on PRs, we introduce Challenge Publish Handshake.

In this two part handshake, the github action first contacts
nyrkio.com claiming to be the author of a currently running workflow run.
Nyrkiö then verifies from Github API that such a run id + author username
actually exist. It then returns a random string - the public challenge -
to the connecting client. The client creates and uploads a workload
artifact that contains this random string. For convenience also
the artifact name shall include the string. This proves to
Nyrkiö that the client that is at the other end of the connection
is also in control of the workflow.

In addition to the PR use case, this can just in general be used to
allow someone to use nyrkio/change-detection without needing to
first create a user account and token on nyrkio.com